CLOSED BETA UPDATE - 337 "64 bit gold"

+++ VikingEngine/DSSWars/Interface/PlayerHud_Head.cs
@@ -82,8 +82,8 @@ namespace VikingEngine.DSSWars.Interface
         {
             //LocalPlayer localPlayer = player.GetLocalPlayer();
 
-            int gold;
-            int income;
+            long gold;
+            long income;
 
             int workForce;
             int totalStrength;
